+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add static OAuth client for debugging
 - Add GitHub Actions and Dependabot configuration
 - Add Maven plugins and repository definitions used by workflows
+- Add interface to manage clients (RFS-254)
 
 ### Changed